+++ generators/project/templates/devtools/tasks/+workspace_createWebPackage.js
@@ -1,0 +1,6 @@
+module.exports = function(grunt) {
+    'use strict';
+    grunt.registerTask('+createWebpackage', 'Create a new WebPackage.', [
+        'prompt:createWebPackage','default'
+    ]);
+};