+++ generators/project/templates/devtools/tasks/configs/availabletasks.js
@@ -1,0 +1,19 @@
+/*jshint node: true */
+/*global module */
+'use strict';
+
+module.exports.tasks = {
+    availabletasks: {
+        tasks: {
+            options: {
+                filter: 'include',
+                tasks: [
+                    '+createWebpackage',
+                    '+startWebserver',
+                    '+webpackage-upload',
+                    '+webpackage-validateManifestFile'],
+                sort: true
+            }
+        }
+    }
+};